refact(dbs-fk): extract helpers
